+++ src/main/java/plugily/projects/murdermystery/api/StatsStorage.java
@@ -120,7 +120,7 @@ public class StatsStorage {
     CONTRIBUTION_DETECTIVE("contribdetective", true),
     CONTRIBUTION_MURDERER("contribmurderer", true), DEATHS("deaths", true), GAMES_PLAYED("gamesplayed", true), HIGHEST_SCORE("highestscore", true),
     KILLS("kills", true), LOSES("loses", true), WINS("wins", true), LOCAL_CURRENT_PRAY("local_pray", false), LOCAL_GOLD("gold", false), LOCAL_KILLS("local_kills", false),
-    LOCAL_PRAISES("local_praises", false), LOCAL_SCORE("local_score", false);
+    LOCAL_PRAISES("local_praises", false), LOCAL_SCORE("local_score", false), MURDERER_PASS("murderer_pass", true), DETECTIVE_PASS("murderer_pass", true);
 
     private final String name;
     private final boolean persistent;