+++ src/main/java/plugily/projects/murdermystery/user/UserManager.java
@@ -72,7 +72,7 @@ public class UserManager {
   public List<User> getUsers(Arena arena) {
     List<User> list = new ArrayList<>();
 
-    for (Player player : arena.getPlayers()) {
+    for(Player player : arena.getPlayers()) {
       list.add(getUser(player));
     }
 
@@ -84,17 +84,17 @@ public class UserManager {
       return;
     }
     //apply before save
-    fixContirbutionStatistics(user);
+    fixContributionStat(user);
     database.saveStatistic(user, stat);
   }
 
   public void loadStatistics(User user) {
     database.loadStatistics(user);
     //apply after load to override
-    fixContirbutionStatistics(user);
+    fixContributionStat(user);
   }
 
-  private void fixContirbutionStatistics(User user) {
+  private void fixContributionStat(User user) {
     if(user.getStat(StatsStorage.StatisticType.CONTRIBUTION_DETECTIVE) <= 0) {
       user.setStat(StatsStorage.StatisticType.CONTRIBUTION_DETECTIVE, 1);
     }